+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - support for text map values (added in SDK 0.11.32)
 - support for CreateAndExercise command (added in SDK 0.12.9)
+- support for GetFlatTransactionById and GetFlatTransactionByEventId (added in SDK 0.12.14)
+
+### Fixed
+- add validation to GetTransactionById endpoint
 
 ## [0.5.0]
 ### Changed